+++ Aesthetics.DTO.NetCore/DataObject/InvoiceDetail.cs
@@ -16,9 +16,9 @@ namespace Aesthetics.DTO.NetCore.DataObject
         public string UserName { get; set; }
         public int ProductID { get; set; }
         public string ProductName { get; set; }
-        public float SellingPrice { get; set; }
+        public double SellingPrice { get; set; }
         public int TotalQuantity { get; set; }
-        public float TotalMoney { get; set; }
+        public double TotalMoney { get; set; }
         public int DeleteStatus { get; set; }
         public string Type { get; set; }
         public Invoice Invoice { get; set; }
